+++ server/src/modules/request/requestRepository.ts
@@ -1,5 +1,5 @@
 import databaseClient from "../../../database/client";
-import type { Rows } from "../../../database/client";
+import type { Result, Rows } from "../../../database/client";
 
 interface Request {
   id: number;
@@ -17,6 +17,17 @@ class RequestRepository {
     );
 
     return rows as Request[];
+  }
+
+  async update(request: Request) {
+    // Execute the SQL UPDATE query to update an existing category in the "category" table
+    const [result] = await databaseClient.query<Result>(
+      "update request set date = ?, title = ?, theme = ?, details = ?, where id = ?",
+      [request.date, request.title, request.theme, request.details, request.id],
+    );
+
+    // Return how many rows were affected
+    return result.affectedRows;
   }
 }
 export default new RequestRepository();
